Removed dropdown content
Realized I was making this thing too complex. I just need those top buttons to link to the future page components. I don't need the specific projects in the dropdown when I can just put that in the sidebar.

So removing the dropdown items since that was complicating the navigation and all that.