Added some images and text to the placeholder site
Added in more pictures so I can actually see the difference when I click on different categories. Also added in the option to the content pane for having text.

I will definitely need to actually plan out the structure of this site though. Don't know what I'm going to put on here.